move funcs to modules
