+++ src/super_gradients/common/environment/monitoring/cpu.py
@@ -1,15 +1,5 @@
 import psutil
 
-from super_gradients.common.environment.monitoring.utils import bytes_to_megabytes
 
-
-def get_disk_usage_percent() -> float:
-    return psutil.disk_usage("/").percent
-
-
-def get_io_read_mbs() -> float:
-    return bytes_to_megabytes(psutil.disk_io_counters().read_bytes)
-
-
-def get_io_write_mbs() -> float:
-    return bytes_to_megabytes(psutil.disk_io_counters().write_bytes)
+def get_cpu_percent():
+    return psutil.cpu_percent(interval=None, percpu=False)
