+++ src/super_gradients/common/environment/monitoring/gpu/__init__.py
@@ -1,4 +1,5 @@
 from super_gradients.common.environment.monitoring.gpu.gpu import (
+    init_nvidia_management_lib,
     count_gpus,
     get_device_memory_utilization,
     get_device_memory_allocated_percent,
@@ -9,6 +10,7 @@ from super_gradients.common.environment.monitoring.gpu.gpu import (
 )
 
 __all__ = [
+    "init_nvidia_management_lib",
     "count_gpus",
     "get_device_memory_utilization",
     "get_device_memory_allocated_percent",
