+++ src/super_gradients/common/environment/monitoring/virtual_memory.py
@@ -1,0 +1,5 @@
+import psutil
+
+
+def virtual_memory_used_percent():
+    return psutil.virtual_memory().percent
